+++ src/components/layouts/Ranking/Children/Global/usersData.ts
@@ -1,6 +1,6 @@
 import avatarImg from "@/assets/avatars/avatar.png";
 
-// Tempporary data for testing tables
+// Temporary data for testing tables
 export interface User {
     rank: number;
     username: string;
