+++ src/components/ui/custom/shared/Ranking/PlayerRow/PlayerRow.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { TableRow, TableCell } from "@/components/ui/shadcn/table";
 import ReactCountryFlag from "react-country-flag";
 import { useNavigate } from "react-router-dom";
-import { User } from "./usersData";
+import { User } from "../../../../../layouts/Ranking/Children/Global/usersData";
 
 interface Props {
     player: User;
