+++ src/components/ui/custom/shared/Ranking/SortButtons/useSortPlayers.ts
@@ -1,5 +1,5 @@
 import { useMemo, useState } from "react";
-import { User } from "@/components/layouts/Ranking/Children/Global/GlobalTable/usersData";
+import { User } from "@/components/layouts/Ranking/Children/Global/usersData";
 
 export type SortKey = "records" | "hardest" | "gold";
